+++ src/main/java/com/seniors/common/exception/ExceptionRestControllerAdvice.java
@@ -1,27 +1,34 @@
 package com.seniors.common.exception;
 
-import com.seniors.common.dto.ErrorResponse;
 import com.seniors.common.constant.ResultCode;
+import com.seniors.common.dto.ErrorResponse;
+import com.seniors.common.dto.ResponseDto;
 import com.seniors.common.exception.type.BadRequestException;
 import com.seniors.common.exception.type.InternalException;
 import com.seniors.common.exception.type.NotFoundException;
+import com.seniors.common.exception.type.ViewException;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindException;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import static org.springframework.web.reactive.function.server.ServerResponse.badRequest;
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.seniors.common.constant.ResultCode.BAD_REQUEST;
 
 @Slf4j
-@RestControllerAdvice(annotations = RestController.class)
+@RestControllerAdvice
 public class ExceptionRestControllerAdvice extends ResponseEntityExceptionHandler {
 	@ExceptionHandler
 	public ResponseEntity<Object> validation(ConstraintViolationException e, WebRequest request) {
@@ -48,11 +55,28 @@ public class ExceptionRestControllerAdvice extends ResponseEntityExceptionHandle
 		return handleExceptionInternal(e, ResultCode.NOT_FOUND, request);
 	}
 
-//	@Override
-//	protected ResponseEntity<Object> handleExceptionInternal(Exception ex, Object body,
-//	                                                         HttpHeaders headers, HttpStatus status, WebRequest request) {
-//		return handleExceptionInternal(ex,  ResultCode.valueOf(status), headers, status, request);
-//	}
+	@ExceptionHandler(ViewException.class)
+	@ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
+	public ModelAndView viewException(ViewException e) {
+		log.error("Unknown Page Error.", e);
+		ModelAndView mv = new ModelAndView("error/500");
+		mv.addObject("errorMessage", e.getMessage());
+		return mv;
+	}
+
+	@ExceptionHandler(BindException.class)
+	@ResponseStatus(HttpStatus.BAD_REQUEST)
+	public ResponseDto validation(BindException e) {
+		BindingResult bindingResult = e.getBindingResult();
+		List<ObjectError> objectErrors = bindingResult.getAllErrors();
+		List<String> errorMessages = new ArrayList<>();
+
+		for (ObjectError objectError : objectErrors) {
+			errorMessages.add(objectError.getDefaultMessage());
+		}
+		String errorMessage = String.join(", ", errorMessages);
+		return ResponseDto.of(false, BAD_REQUEST, errorMessage);
+	}
 
 
 	private ResponseEntity<Object> handleExceptionInternal(Exception e, ResultCode errorCode,
